+++ README.md
@@ -30,14 +30,14 @@ problems, and they can continue working as usual.
 
 ## Sidecar
 
-![sidecar check](/docs/visio/1.%20Sidecar%20check.PNG)
+![sidecar check](/_docs/visio/1.%20Sidecar%20check.PNG)
 
 The sidecar will check the status about the `GitLab Runner` and export a
 [liveness health check][liveness health check].
 
 [liveness health check]: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-a-liveness-command
 
-![sidecar check](/docs/visio/1.%20Sidecar%20check%20explained.PNG)
+![sidecar check](/_docs/visio/1.%20Sidecar%20check%20explained.PNG)
 
 ## Decider
 
@@ -45,7 +45,7 @@ The decider has 2 jobs:
 
 ### Status Check
 
-![sidecar check](/docs/visio/2.%20Decider%20check.PNG)
+![sidecar check](/_docs/visio/2.%20Decider%20check.PNG)
 
 - Checking the `liveness health check` of each available `GitLab Runner`
 installed on the same cluster
@@ -53,11 +53,11 @@ installed on the same cluster
 - Exporting the statuses (`Online / Total`) to the configured
 `GitLab` project
 
-![sidecar check](/docs/visio/2.%20Decider%20check%20explained.PNG)
+![sidecar check](/_docs/visio/2.%20Decider%20check%20explained.PNG)
 
 ### Update Runners
 
-![sidecar check](/docs/visio/3.%20Deciders%20action.PNG)
+![sidecar check](/_docs/visio/3.%20Deciders%20action.PNG)
 
 - Read the statuses of both clusters
 
@@ -70,7 +70,7 @@ If stronger, it will update the tag list of the `GitLab Runners`
 to the configured desired main cluster tag so users will not have to change
 their `.gitlab-ci.yml` files.
 
-![sidecar check](/docs/visio/3.%20Deciders%20action%20-%20win.PNG)
+![sidecar check](/_docs/visio/3.%20Deciders%20action%20-%20win.PNG)
 
 #### Weaker
 
@@ -78,4 +78,4 @@ If weaker, it will remove the configured desired main cluster tag from
 the tag list so users will stop using those runners without updating their
 `.gitlab-ci.yml` file.
 
-![sidecar check](/docs/visio/3.%20Deciders%20action%20-%20lose.PNG)
+![sidecar check](/_docs/visio/3.%20Deciders%20action%20-%20lose.PNG)
